src/utils/configure_hugo.py
import os
import yaml

# Function to configure a Hugo theme with template tags and Decap CMS integration
def configure_hugo(path):
    print(f"Configuring Hugo theme at {path}...")
    
    # Analyze the theme structure
    analyze_theme_structure(path)

    # Apply Hugo template tags
    apply_hugo_tags(path)

    # Configure data types
    configure_data_files(path)

    # Integrate with Decap CMS
    integrate_decap_cms(path)

    return "Hugo theme configuration complete"

# Analyze the theme structure
def analyze_theme_structure(path):
    print("Analyzing theme structure...")
    # Placeholder for analysis logic

# Apply Hugo template tags
def apply_hugo_tags(path):
    print("Applying Hugo template tags...")
    # Placeholder for tag application logic

# Configure data files
def configure_data_files(path):
    print("Configuring data files...")
    # Placeholder for data configuration logic

# Integrate with Decap CMS
def integrate_decap_cms(path):
    print("Integrating with Decap CMS...")
    # Placeholder for Decap CMS integration logic 