requirements.txt
openai
flask
pyyaml
openai 
beautifulsoup4